requirements.txt
Mastodon.py
python-mpd2
